Made group tabs load sooner
I don't really know what the deal is with TabView._initFrame, but calling it earlier apparently makes it set up the groups data earlier, allowing the extension to load the group tabs BEFORE trying to load the active page.  It doesn't seem like it should work this way -- it seems like TabView._initFrame is calling its callbacks at the wrong time -- but anyway, it appears to work.