Simpler approach to storing sort order
Instead of a dict mapping group ID to sort position, just use a list of the group IDs.  The order of the list is the sort order.

Note: I also tried just storing the sort order on the groups themselves (inside GroupItems.groupItems), but that doesn't work, because the information isn't saved/restored with the browser session.