Added ability to bookmark a group (via group tab context menu)
Creates a new bookmark group and adds a bookmark for each tab in the group.

Unfortunately this is kind of slow.  It sometimes takes up to 100ms per tab, which seems pretty nuts.  The slowness seems to be inherent in the actual bookmarking API calls, though, so I can't see any way around it.  It may be faster if using a Firefox session that isn't super huge with all kinds of stored data.